refactor: move rotas protegidas para rota app
