+++ src/app/app/dashboard/layout.tsx
@@ -1,0 +1,3 @@
+export default function Layout({ children }: { children: React.ReactNode }) {
+	return <>{children}</>;
+}
