+++ src/app/app/dashboard/traceability/[id]/page.tsx
