+++ src/app/app/layout.tsx
@@ -1,0 +1,20 @@
+import { SidebarInset, SidebarProvider } from '@/components';
+import SessionProvider from '@/features/core/providers/sessionProvider';
+import { AppSidebar } from '@/features/dashboard';
+
+export default async function Layout({
+	children,
+}: {
+	children: React.ReactNode;
+}) {
+	return (
+		<div>
+			<SessionProvider>
+				<SidebarProvider>
+					<AppSidebar />
+					<SidebarInset>{children}</SidebarInset>
+				</SidebarProvider>
+			</SessionProvider>
+		</div>
+	);
+}
